feat: created carousel.js for the carousel code, imported it to singleListing
